cli: move common functions to util file
